refactor: move remove reserved keys
