+++ sdk/src/androidTest/java/ly/count/android/sdk/UtilsInternalLimitsTests.java
@@ -208,4 +208,35 @@ public class UtilsInternalLimitsTests {
         //Assert.assertEquals("2", values.get("a2"));
         //Assert.assertEquals("4", values.get("a4"));
     }
+
+    @Test
+    public void removeReservedKeysFromSegmentation() {
+        Map<String, Object> values = new HashMap<>();
+
+        UtilsInternalLimits.removeReservedKeysFromSegmentation(values, new String[] {}, "", mock(ModuleLog.class));
+        Assert.assertEquals(0, values.size());
+
+        UtilsInternalLimits.removeReservedKeysFromSegmentation(values, new String[] { "a", "", null }, "", mock(ModuleLog.class));
+        Assert.assertEquals(0, values.size());
+
+        values.put("b", 1);
+        Assert.assertEquals(1, values.size());
+        UtilsInternalLimits.removeReservedKeysFromSegmentation(values, new String[] { "a", "a1", "", null }, "", mock(ModuleLog.class));
+        Assert.assertEquals(1, values.size());
+        Assert.assertTrue(values.containsKey("b"));
+
+        values.put("a", 2);
+        Assert.assertEquals(2, values.size());
+        UtilsInternalLimits.removeReservedKeysFromSegmentation(values, new String[] { "a", "a1", "", null }, "", mock(ModuleLog.class));
+        Assert.assertEquals(1, values.size());
+        Assert.assertTrue(values.containsKey("b"));
+
+        values.put("a", 2);
+        values.put("c", 3);
+        Assert.assertEquals(3, values.size());
+        UtilsInternalLimits.removeReservedKeysFromSegmentation(values, new String[] { "a", "a1", "", null }, "", mock(ModuleLog.class));
+        Assert.assertEquals(2, values.size());
+        Assert.assertTrue(values.containsKey("b"));
+        Assert.assertTrue(values.containsKey("c"));
+    }
 }
